add sub-directories to organize the different object types in the gem
