+++ lib/shopify_app.rb
@@ -10,22 +10,24 @@ require 'shopify_app/configuration'
 # engine
 require 'shopify_app/engine'
 
-# helpers and concerns
-require 'shopify_app/session_storage'
-require 'shopify_app/localization'
-require 'shopify_app/login_protection'
-require 'shopify_app/webhook_verification'
-require 'shopify_app/app_proxy_verification'
+# utils
 require 'shopify_app/utils'
 
+# controller concerns
+require 'shopify_app/controller_concerns/localization'
+require 'shopify_app/controller_concerns/login_protection'
+require 'shopify_app/controller_concerns/webhook_verification'
+require 'shopify_app/controller_concerns/app_proxy_verification'
+
 # jobs
-require 'shopify_app/webhooks_manager_job'
-require 'shopify_app/scripttags_manager_job'
+require 'shopify_app/jobs/webhooks_manager_job'
+require 'shopify_app/jobs/scripttags_manager_job'
 
 # managers
-require 'shopify_app/webhooks_manager'
-require 'shopify_app/scripttags_manager'
+require 'shopify_app/managers/webhooks_manager'
+require 'shopify_app/managers/scripttags_manager'
 
-# session repository
-require 'shopify_app/shopify_session_repository'
-require 'shopify_app/in_memory_session_store'
+# session
+require 'shopify_app/session/session_storage'
+require 'shopify_app/session/session_repository'
+require 'shopify_app/session/in_memory_session_store'
