+++ lib/shopify_app/controller_concerns/app_proxy_verification.rb
