+++ lib/shopify_app/controller_concerns/localization.rb
