+++ lib/shopify_app/controller_concerns/login_protection.rb
