+++ lib/shopify_app/jobs/scripttags_manager_job.rb
