+++ lib/shopify_app/managers/scripttags_manager.rb
