+++ lib/shopify_app/session/session_repository.rb
