+++ lib/shopify_app/session/session_storage.rb
